refactor(experiment): move make_experiment_output_dir to experiment.py; enforce UUID suffix
Move EXP_DIR and make_experiment_output_dir out of __init__.py into experiment.py,
which owns the Experiment class and evaluation_id logic. Add a model_validator that
appends a _{uuid8} suffix to output_dir.name if absent, so evaluation_id is always
globally unique. __init__.py is now a thin re-export. The tmp_dir fixture gains a
UUID-suffixed subdir to satisfy the validator in tests; the save_config test is
updated to derive the expected path from exp.output_dir rather than hardcoding it.

Adds base_dir parameter to make_experiment_output_dir for easier testing.
